+++ chain/middleware_stack_test.go
@@ -1,4 +1,4 @@
-package velocity
+package chain
 
 import (
 	"testing"